backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
Pillow==10.1.0
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==23.2.1

# AI/ML Dependencies for ControlNet + Stable Diffusion
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.21.0
transformers>=4.25.0
accelerate>=0.20.0
controlnet-aux>=0.4.0
opencv-python>=4.8.0
numpy>=1.24.0
scipy>=1.10.0
xformers>=0.0.20  # Optional: for memory optimization